monopoly_bot/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
aiofiles==23.2.1
asyncpg>=0.29.0
python-dotenv==1.0.1
